refactor: move header and footer to new location
