+++ src/components/layout/index.ts
@@ -1,1 +1,3 @@
 export { default as Container } from './Container';
+export { default as Header } from './Header';
+export { default as Footer } from './Footer';
